flask/strands/requirements.txt
Flask>=0.12.2
biopython>=1.70
